changed according to comments
