+++ androidsdk/src/main/java/com/wootric/androidsdk/Wootric.java
@@ -12,8 +12,7 @@ import com.wootric.androidsdk.utils.PreferencesUtils;
 /**
  * Created by maciejwitowski on 4/10/15.
  */
-public class Wootric implements
-        GetAccessTokenTask.OnAccessTokenReceivedListener{
+public class Wootric {
 
     private final Activity activity;
     UserManager userManager;
@@ -38,20 +37,7 @@ public class Wootric implements
 
         User user = new User(clientId, clientSecret, accountToken);
         userManager = new UserManager(activity, user, connectionUtils, preferencesUtils);
-        if(shouldResendUnsentResponse()) {
-            new GetAccessTokenTask(user, this, connectionUtils).execute();
-        }
         return userManager;
-    }
-
-    @Override
-    public void onAccessTokenReceived(String accessToken) {
-        CreateResponseTask.restartIfUnsentMessageExist(activity,ConnectionUtils.get(),accessToken);
-    }
-
-    private boolean shouldResendUnsentResponse(){
-        PreferencesUtils preferencesUtils = PreferencesUtils.getInstance(activity);
-        return preferencesUtils.getUnsentResponse()!=null && !(preferencesUtils.getAccessToken() != null && preferencesUtils.getEndUserId() != Constants.INVALID_ID);
     }
 
     private Wootric(Activity activity) {
