+++ androidsdk/src/main/java/com/wootric/androidsdk/utils/PreferencesUtils.java
@@ -3,6 +3,11 @@ package com.wootric.androidsdk.utils;
 import android.content.Context;
 import android.content.SharedPreferences;
 
+import com.wootric.androidsdk.objects.EndUser;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
 import java.util.Date;
 
 import static com.wootric.androidsdk.utils.Constants.INVALID_ID;
@@ -82,10 +87,23 @@ public class PreferencesUtils {
         return prefs().getString(KEY_PREVIOUS_FAILED_RESPONSE, null);
     }
 
-    public void setUnsent_response(String unsent_response) {
-        prefs().edit().putString(KEY_PREVIOUS_FAILED_RESPONSE, unsent_response).apply();
+    public void saveUnsentResponse(EndUser endUser, String text, int score, String originUrl) {
+        String json = "";
+        try {
+            JSONObject obj = new JSONObject();
+            obj.put("endUser", endUser.toJson());
+            obj.put("originUrl", originUrl);
+            obj.put("score", score);
+            obj.put("text", text);
+            json = obj.toString();
+        } catch (JSONException e){
+            e.printStackTrace();
+        }
+        prefs().edit().putString(KEY_PREVIOUS_FAILED_RESPONSE, json).apply();
     }
-
+    public void clearUnsentResponse(){
+        prefs().edit().putString(KEY_PREVIOUS_FAILED_RESPONSE, null).apply();
+    }
     public void setLastSurveyed(long lastSurveyed) {
         prefs().edit().putLong(KEY_LAST_SURVEYED, lastSurveyed).apply();
     }
